% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert Genotype}
\usage{
convert(
  input_geno,
  type = c("hmp1", "hmp2", "num", "vcf"),
  missingrate = 0.2,
  maf = 0.05,
  impute = TRUE,
  dir = NULL
)
}
\arguments{
\item{input_geno}{genotype in HapMap format,VCF format or in numeric format. The names of individuals should be provided. Missing (NA) values are allowed.}

\item{type}{the type of genotype. There are three options: "hmp1" for genotypes in HapMap format with single bit, "hmp2" for genotypes in HapMap format with double bit, "vcf" for genotypes in VCF and "num" for genotypes in numeric format.}

\item{missingrate}{max missing percentage for each SNP, default is 0.2.}

\item{maf}{minor allele frequency for each SNP, default is 0.05.}

\item{impute}{logical. If TRUE, imputation. Default is TRUE.}

\item{dir}{This file is for the input path.When type= "vcf", it must be provided, default is NULL.}
}
\value{
A matrix of genotypes in numeric format, coded as 1, 0, -1 for AA, Aa, aa. Each row represents an individual and each column represents a marker. The rownames of the matrix are the names of individuals.
}
\description{
Convert genotypes in HapMap format ,VCF format or in numeric format for convert function.
}
\examples{
## load genotype in HapMap format with double bit
data(input_geno)

## convert genotype for Maize6KGSP package
inbred_gen <- convert(input_geno, type = "hmp2")


## load genotype in numeric format
data(input_geno1)
head(input_geno1)

## convert genotype for Maize6KGSP package
inbred_gen1 <- convert(input_geno1, type = "num")

}
