% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Evaluate Trait Predictability via Cross Validation}
\usage{
cv(
  fix = NULL,
  gen,
  y,
  method = "GBLUP",
  drawplot = TRUE,
  nfold = 5,
  nTimes = 1,
  seed = 1234,
  CPU = 1
)
}
\arguments{
\item{fix}{a design matrix of the fixed effects.}

\item{gen}{a matrix (n x m) of genotypes.}

\item{y}{a vector(n x 1) of the phenotypic values.}

\item{method}{six GS methods including 'GBLUP', 'BayesB', 'RKHS', 'PLS', 'LASSO', 'XGBoost'.
Users may select one of these methods or all of them simultaneously with 'ALL'. Default is 'GBLUP'.}

\item{drawplot}{when method ='ALL', user may select TRUE for a barplot about six GS methods. Default is TRUE.}

\item{nfold}{the number of folds. Default is 5.}

\item{nTimes}{the number of independent replicates for the cross-validation. Default is 1.}

\item{seed}{the random number. Default is 1234.}

\item{CPU}{the number of CPU.}
}
\value{
Trait predictability
}
\description{
The cv function evaluates trait predictability based on six GS methods via k-fold cross validation. The trait predictability is defined as the squared Pearson correlation coefficient between the observed and the predicted trait values.
}
\examples{
\donttest{
## load example data
data(hybrid_phe)
data(input_geno)

## convert original genotype
inbred_gen <- convert(input_geno, type = 'hmp2')

##additive model infer the additive genotypes of hybrids
gena <- infergen(inbred_gen, hybrid_phe)$add

##additive model
R2<-cv(fix=NULL,gen=gena,y=hybrid_phe[,3],method ='GBLUP',nfold=5,nTimes=1,seed=1234,CPU=1)}
}
