% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_int.R
\name{geno_int}
\alias{geno_int}
\title{Genotype data integration}
\usage{
geno_int(
  train_input_geno,
  breed_input_geno,
  type,
  missingrate,
  maf,
  impute = TRUE,
  dir = NULL,
  train_crosses = NULL,
  breed_crosses = NULL
)
}
\arguments{
\item{train_input_geno}{the genotypic data of the training population in HapMap format,VCF format or in numeric format. The names of individuals should be provided. Missing (NA) values are allowed.}

\item{breed_input_geno}{the genotypic data of the breeding population in HapMap format,VCF format or in numeric format. The names of individuals should be provided. Missing (NA) values are allowed.}

\item{type}{the type of genotype. There are three options: 'hmp1' for genotypes in HapMap format with single bit, 'hmp2' for genotypes in HapMap format with double bit, 'vcf' for genotypes in VCF and 'num' for genotypes in numeric format.}

\item{missingrate}{max missing percentage for each SNP, default is 0.2.}

\item{maf}{minor allele frequency for each SNP, default is 0.05.}

\item{impute}{logical. If TRUE, imputation. Default is TRUE.}

\item{dir}{This file is for the input path.When type= 'vcf', it must be provided, default is NULL.}

\item{train_crosses}{a data frame with two columns. The first column and the second column are the names of male and female parents of the corresponding  training population crosses. Default is NULL.}

\item{breed_crosses}{a data frame with two columns. The first column and the second column are the names of male and female parents of the corresponding  breed population crosses. Default is NULL.}
}
\value{
A matrix of genotypes in numeric format, coded as 1, 0, -1 for AA, Aa, aa. Each row represents an individual and each column represents a marker. The rownames of the matrix are the names of individuals.
}
\description{
Match the genotypic data of the training population and breeding population, ensuring consistency of markers.
}
