% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_predict.R
\name{phenotype_predict}
\alias{phenotype_predict}
\title{Predict the Performance of individuals}
\usage{
phenotype_predict(
  train_fix = NULL,
  breed_fix = NULL,
  train_geno,
  breed_geno,
  train_phe,
  method = "GBLUP"
)
}
\arguments{
\item{train_fix}{a matrix of fixed effect for training.}

\item{breed_fix}{a matrix of fixed effect for breeding.}

\item{train_geno}{a matrix (n x m) of training population genotypes.}

\item{breed_geno}{a matrix (n x m) of breddin population genotypes..}

\item{train_phe}{a vector (n x 1) of phenotype for the training population phenotypes.}

\item{method}{six GS methods including 'GBLUP', 'BayesB', 'RKHS', 'PLS', 'LASSO', 'XGBoost'. Users may select one of these methods. Default is 'GBLUP'.}
}
\value{
a phenotype of breeding population.
}
\description{
The phenotype and genotype data of the training population were used to construct models to predict the phenotype of the breeding population.
}
\examples{
\donttest{
## load example data from Maize6KGSPred package
data(hybrid_phe)
data(input_geno)

## convert original genotype
inbred_gen <- convert(input_geno, type = 'hmp2')

##additive model infer the additive and dominance genotypes of hybrids
gena <- infergen(inbred_gen, hybrid_phe)$add
set.seed(123)
index <- sample(c(TRUE, FALSE), nrow(gena), replace=TRUE, prob=c(0.8,0.2))
train <- gena[index, ]
breed <- gena[!index, ]
t_phe <-hybrid_phe[index, 3]
p <- phenotype_predict(train_geno = train, breed_geno = breed, train_phe = t_phe)
 }
}
